% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{launch_shinybisonpic}
\alias{launch_shinybisonpic}
\title{Run shinybisonpic App}
\usage{
launch_shinybisonpic()
}
\description{
This function will launch the app locally.
}
\examples{
\dontrun{
launch_shinybisonpic()
}
}
