% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybisonpic-package.R
\docType{package}
\name{shinybisonpic-package}
\alias{shinybisonpic}
\alias{shinybisonpic-package}
\title{shinybisonpic: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

Other contributors:
\itemize{
  \item Alberta Environment and Parks [copyright holder]
}

}
\keyword{internal}
